+++ webapps/controllers/dm/News_controller.php
@@ -270,7 +270,9 @@ class News_controller extends CI_Controller
     public function delete_news_category()
     {
         $news = $this->News_model->getNewsByCatId($this->uri->segment(3));
-        unlink('./' . $news['img_src']);
+        if(!is_null($news['img_src'])) {
+            unlink('./' . $news['img_src']);
+        }
         $this->News_model->delete($this->uri->segment(3));
         redirect('manage-study-news', 'refresh');
     }
