+++ webapps/controllers/dm/Schedule_controller.php
@@ -161,7 +161,9 @@ class Schedule_controller extends CI_Controller
     public function delete_schedule()
     {
         $news = $this->News_model->getNewsById($this->uri->segment(3));
-        unlink('./' . $news['img_src']);
+        if(!is_null($news['img_src'])) {
+            unlink('./' . $news['img_src']);
+        }
         $this->News_model->delete($this->uri->segment(3));
         redirect('schedule-manager', 'refresh');
     }
